refactor: move functions to separate files
Refactored preventScrolling into own function and renamed freezeScroll
to preventScrolling, moved both to functions folder.

Moved addScore to functions folder.
