+++ src/functions/index.js
@@ -5,3 +5,6 @@ export * from './isAtRightEdge'
 
 export * from './drawTetromino'
 export * from './undrawTetromino'
+
+export * from './preventScrolling'
+export * from './addScore'
